Move new functions to different files
